Adding additional error context to Pagerduty Notification
We wanted to be able to include more specific build job errors in
Pagerduty notifications. To achieve this, we wrote some Python to
collect build log output and strip away any successful tasks.
Anything failing is then returned as a build log summary to a bash
script which does the hand off to Pagerduty.
Pagerduty also reports the Concourse url of the failing job for
convenience.
An incident key of Concourse pipeline name and build job is used to
ensure the same fail event in subsequent job runs is not triggering
Pagerduty endlessly.
concourse-smuggler-resource is used as the basis for this resource
as it already conveniently includes many useful build variables in
its execution context.

Signed-off-by: Hector Rivas Gandara <hector.gandara@fil.com>