Handle inevitable ReadTimeoutError when event stream exhausted
When all logs have been shuttled through the event stream, this resource
will hold open the connection, but a read timeout will inevitably be
reached when no further log output is sent. We should handle this as a
way of neatly breaking out of the event stream loop, but also if there
are any genuine ReadTimeoutErrors for any other reason.

We are also outputting an EventReaderWaterMark message on stderr so
that we can filter any log output that this resource generates on the
event stream as this is probably just noise we don't want.

Signed-off-by: Hector Rivas Gandara <hector.gandara@fil.com>