Allow users to specify the event type and incident_key
This adds a new param for the event `action`, which controls
whether an alert will be triggered, resolved, or acknowledged
based on the event. Should allow users to create `on_success` events
which auto-resolve incidents that were created by the pipeline's
`on_failure` event.

This also fixes a bug where the optional incident_key param was
ignored, defaulting it to what used to be forced on the event.

Also fixed are issues where problems connecting to pagerduty would
trigger non-json output from the resource, resulting in Concourse errors

Signed-off-by: Ed Payne <epay0@allstate.com>